backend/requirements.txt
Django==4.0.2
djangorestframework==3.13.1
djoser==2.1.0
django-cors-headers==3.11.0
drf-yasg==1.20.0
django-filter==21.1
Pillow==9.0.1
psycopg2-binary






